[GStreamer] Fix video resize problems after quality change
Quality changes trigger flush operations, which trigger DRAIN queries,
which clear the m_sample in the player private base. With no m_sample,
there's no way to get the video size, so the default size is used. This
causes resizing problems (sudden shrinks/stretches).

This patch removes the expensive part of the m_sample (the GstBuffer)
but keeps the caps in a new dummy sample. In this way, the original
video size can still be known and the resizing problems disappear.