[GStreamer] Fixed a CRITICAL in GstVideoFrameHolder
by avoiding holding the frame when info and meta dimensions aren't the same.
That would cause a CRITICAL in gst_video_frame_map().

This can happen on prerolling (eg: on PLAYING->PAUSED) after a quality change,
because the first frame (old quality) is returned but the meta of the high
quality one is kept.