[EME] Address concurrenct access to session map.
This map was being corrupted by concurrent data accesses with the gstreamer
decryptor elements. The result of that was slient segfaults in the web process,
and the only smoking gun being the severed IPC connection to the framework.

The reason std::map with std::string is used instead of HashMap is
because HashMap does not work with Ref or RefPtr as part of the type for
it's hashing functions, and std::map does not work with WTF String
because it doesn't have the right operator traits.

* platform/graphics/gstreamer/eme/CDMOpenCDM.cpp:
(WebCore::CDMInstanceOpenCDM::requestLicense):
(WebCore::CDMInstanceOpenCDM::updateLicense):
(WebCore::CDMInstanceOpenCDM::loadSession):
(WebCore::CDMInstanceOpenCDM::removeSessionData):
(WebCore::CDMInstanceOpenCDM::closeSession):
(WebCore::CDMInstanceOpenCDM::sessionIdByInitData const):
(WebCore::CDMInstanceOpenCDM::isSessionIdUsable const):
* platform/graphics/gstreamer/eme/CDMOpenCDM.h: