[MSE] Don't remove buffers from decodeQueue when evicting
In some situations, a trackBuffer decodeQueue may hold samples but enqueuing
may be retained because the SourceBuffer isn't readyForMoreSamples. If further
appends demand free space to hold the new samples, evictCodedFrames() will try
to make room and delete old buffered samples up to currentTime.

When the latency since the sample is enqueued until it's decoded is too high,
the video decoder can lag behind currentTime and not yet decoded samples can
become eligible for eviction. In that case, samples are not only removed from
the trackBuffer, but also from the (stalled because not readyForMoreSamples)
decodeQueue. When the SourceBuffer becomes readyForMoreSamples again, the
samples aren't available from enqueuing anymore and the video stalls.

This commit distinguishes several kinds of sample deletion scenarios:

1. When the samples are deleted because a new samples want to overwrite them
   (eg: on a quality change).
2. When samples are deleted explicitly by JavaScript.
3. When samples are evicted in order to free memory and make room for new
   samples.

In all the cases the samples were also removed from the decodeQueue (if
present) and the trackBuffer was marked as needsReenqueueing. However, in case
3 that's not actually needed because future appends will likely continue
providing consecutive data.

This patch supports that case and avoids needsReenqueueing in that case, thus
preventing the video stalls.