[NetworkProcess][OOM][RunLoopSourcePriority] Avoid async IO starving timers
If AsyncIONetwork and DiskCacheRead priorities are higher than
MainThreadSharedTimer the timers get starved. This causes the NetworkProcess
to accumulate MB of data instead of handing it down to the WebProcess (done
using a Timer). This eventually causes an Out Of Memory kill on the
NetworkProcess.

This commit levels the three priorities to the same value, while still leaving
DiskCacheWrite with less priority than DiskCacheRead.