[GStreamer][Nexus] Reduce latency on live MPEG2TS streams with realtime bandwidth
Live MPEG2TS streams with realtime bandwidth (data produced at the same rate
it's consumed, eg: from a TV tuner streamer) may take a long time to fill the
GstQueue2 up to 100% on Nexus, because the vidfilter keeps draining the data
from the queue even when in PAUSED. In tose cases the user sees the first
frame (prerolling) but the video doesn't start moving (PLAYING) until ~10s
later, when the vidfilter has saturated and the queue is 100% filled.

This patch allows an early buffer filled condition when some amount of data
has gone trough the GstQueue2. Empirically, 2 MB is the right value that
doesn't cause random losses of the audio stream, while still providing a
reasonable latency. This new behaviour is only applied to live streams.