Update m_isStreaming with caution.
Calling totalBytes() has a side effect of setting the mentioned
variable and once it's set it's never revalidated. The issue seems
to be it can be called very early when duration isn't known and
can not be determined yet but such call will cause the wrong
assumption of duration being 0 (indicating live stream).
Having assumed the live stream has an impact on media disk cache
being enabled or not.