[GStreamer][MSE] Adjust segmentFixerProbe with last known seek time
This prevents issues on PlaybackPipeline::flush() after a seek. Without this,
the compensation applied by segmentFixerProbe assumes that playback has
started from 0, but after a seek that's not the case. Before this patch,
playback after a flush (after seek) could have a delay as long as the target
seek position.