[GStreamer][MSE][EME] Set system_id on AppendPipeline demuxers using a protection event when CDMInstance is attached
It avoids this qtdemux error on CMAF playback:

  gst_qtdemux_configure_protected_caps:<qtdemux0> stream is protected using cenc, but no cenc protection system information has been found

Note that some CMAF JavaScript implementations don't provide a pssh box in the
data appended to the SourceBuffer, so the regular configuration mechanism based
on pssh isn't working. That's why qtdemux errored out in the first place.

When CDMInstance is attached, a "webkit-media-player-synthesized" protection
event with the system_id is distributed to all the AppendPipelines to be cached
and only injected when the demuxer is ready to accept it. That happens on the
first pushNewBuffer() call.