[GSTUTILS] getElement() leak fix.
In one code path (getting sinks from the playbin) the caller is
responsible for freeing the returned element and the other one
(findElement()) does not copy thus no free is needed. In order
to hide that and properly act no matter the path GRefPtr is now
returned from the function and it properly mnages ref count.