[CoordinatedGraphics][holepunch] Invalidate only PlatformLayerProxies that are not reused
This patch fixes a Disney+ issue with <video> using holepunch. When the video
has a subtitles track but subtitles are hidden, the media controls layer is
enabled anyway, has nothing to paint, covers all the video and causes some
kind of change that creates a new GraphicsLayer to appear for the <video>
element (with a new id, taking the place of the old one that rendered the
hole for the video).

Without this patch, the problem could have been worked around by always enabling
the subtitles, or by setting a solid background color style in the <video>
element.

The new TextureMapperLayer of the <video> doesn't have an associated
(TextureMapperPlatformLayer)m_contentLayer and therefore the hole punch is not
painted. The user sees black instead (the default background). This new
TextureMapperLayer should reuse the TextureMapperPlatformLayer of the old
TextureMapperLayer.

This patch detects that the TextureMapperPlatformLayerProxy is going to be
reused by the new <video> TextureMapperLayer and doesn't invalidate the
proxy. This allows the hole punch code in MediaPlayerPrivateGStreamer to keep
pushing buffers (drawing the hole).