[BCM][GStreamer][MSE] Fix stall when starting video playback from the middle ("continue watching")
After an initial seek on Nexus systems sometimes the Broadcom elements can't
provide a valid position after a preroll, so 0 is used as fallback. This
confuses the MSE seek continuation code, who thinks that currentMediaTime() is
0, instead of the target position of the already completed seek.

With this, updateStates() considers that the current position (0, wrong!) is
not buffered and demotes the readyState from HaveEnoughData to HaveMetadata.

This change detects that the position provided by the Broadcom elements is
invalid and, only in that case, uses the old seek target time cached in
m_seekTime, or 0 as last resort fallback. This effectively fixes the issue.