ONEM-21363: calculate the max time loaded base on network stats from webkitwebsrc element
This fix bases on the change from:
https://trac.webkit.org/changeset/243537/webkit

The original fix was aimed to fix the problem in case of HLS streams, while my implementation
fix the problem with the buffering level in case of progressive playback. It works only for
pipeline with WebkitWebSrc element and when the buffering on disk is disabled.