[GStreamer] Send http-headers event on WebKitWebSrc streaming thread
Added an HttpHeadersEventSender class to help sending the "http-headers" custom
downstream sticky event in the streaming thread instead of sending it from the
main thread (which causes deadlocks).

Ideally, this should be done using gst_element_send_event(), but the GstAppSrc
implementation doesn't work properly on GStreamer 1.16. It will work starting
from 1.18. The commits implementing that support aren't trivial to backport,
so this workaround still makes sense.

When CachedResourceStreamingClient::responseReceived() in the application
needs to send the event, it must not be sent immediately. Events can only be
sent after the segment event. That's why HttpHeadersEventSender uses a probe
to wait for a segment event on the appsrc src pad and only then sends the
custom http-headers event configured on responseReceived(). If the segment
event has previously been sent, the http-headers event is set for delivery and
will be sent immediately as soon as a buffer or an event traverses the probe.

A lock is used to prevent concurrent access to m_httpHeadersEvent from the
probe and from setEvent().