[BCM][GStreamer] Correct buffering level with playpump content in Nexus
Playpump can buffer a lot of content on its own, draining GstQueue2 and
resulting in buffering levels that don't reflect reality. This behaviour can
cause slow starts (GstQueue2 takes longer to fill up if playpump is draining
it) or even rebuffering pauses when the truth is that there's actually plenty
of buffered content (in playpump) that isn't being had into account.