[BCM][GStreamer] Correct buffering level with multiqueue content in Nexus
On top of what playpump may already have been buffered in Nexus, it's also
interesting to account the data buffered in multiqueue to correct the buffered
percentage.

This is only possible on GStreamer 1.18 or above. On earlier versions, the
"stats" multiqueue property isn't implemented and there's no way to know how
much data is actually buffered in the multiqueue.

The "stats" property was added in this commit:

https://github.com/GStreamer/gstreamer/commit/74938f07c2a9b3411716fa7595178942c80e20f4