Video was not scaled properly on successive launch of page / app
It looks like the isVisible inside doesn't reflect the actual visibility
state of the element. On exiting from the test page / app for the first
time the isVisible is set to false by HTMLMediaElement but on next
launch platformShow() is not triggered leaving the isVisible false.

The subsequent calls from GStreamerHolePunchClient bails out as
isVisible is false.

The change tries to align isVisible with element's visibility state by
calling platformShow() / platformHide() from MediaPlayer::setVisible()

Also have reordered the statements in setRectangleToVideoSink() to
* decouple visibility state update from existence of videoSink
* updating isVisible before actually checking for bailing out