[MSE][GStreamer] Fixed video freeze after flush for the segment base
Base can't be on stream time, it is playback time. So in the segmentFixerProbe we have to correct it by taking the sink playback time. For that we need to query
the player about the sink and dig if it's a bin. Then, by using the sink's segment, we can calculate the current playback time.