WebSpeech: Use Optional to track error codes instead of a separate enum
`SpeechError` had the same values as `SpeechSynthesisErrorCode` with the
addition of a `None` value, to sinalize whether we had an error code set
or not.

This commit replaces `SpeechError` with the use of
`WTF::Optional<SpeechSynthesisErrorCode>` to remove the duplicated
information from `SpeechError`.

Note: When upstreaming, `WTF::Optional` would be replaced with `std::optional`.