[DEBUG] Dump pipeline on high memory consumption
The goal is to check current-level-bytes and max-size-bytes of appsrc,
queue2 and multiqueue to check if any of those elements are holding
an unexpectedly huge amount of buffers.

Also, dump WebKitWebView need-data/enough-data to check if there's any
kind of mismatch between NetworkProcess and appsrc and the latter is
being fed with buffers without control and taking a lot of space.