Fix OOM caused by excessive buffering in progressive
This commit is basically a continuation of
57bbecb, which did not fully fix the original issue.

webKitWebSrcEnoughData and webKitWebSrcNeedData are used to pause and unpause
the stream fetching procedure ongoing on the network process. Both functions
operate on the same |paused| variable, but webKitWebSrcNeedData is
called on different thread. This is essentially a data race, which may lead to
incorrect flag state. Flow required for issue reproduction goes like this:

Streaming thread           Main thread
       |                        |
webKitWebSrcNeedData() webKitWebSrcEnoughData()
       |                    paused=true
       |               setDefersLoading(true)
       |                        |
   paused=false                 |
       |                        |
       |               webKitWebSrcEnoughData()
       |                    paused=true
       |               setDefersLoading(true)
       |                        |
       |               // Scheduled from call webKitWebSrcNeedData from Streaming thread
       |               setDefersLoading(false)

As a result, |paused=true|, but the network process is not actually paused.
It will keep sending data to the web process, which will push it to appsrc, the
memory used by appsrc will grow with no limit and the process will eventually
cause an out of memory (OOM) error.

The solution is to always use the |paused| flag on Main thread.